TellMe-develop/TellMe-develop/backend/requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
celery
redis
django-cors-headers
boto3
django-storages
drf-yasg
